requirements.txt
httpx
playwright
python-dotenv
psycopg2-binary
scikit-learn
pandas
numpy
streamlit
folium
streamlit-folium
plotly
schedule
geopy
fake-useragent
joblib
